# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para la gestión de la clínica veterinaria: autenticación, usuarios, clientes, empleados, mascotas, citas, pagos y notificaciones.",
        "title": "API Clínica Veterinaria",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Equipo de Soporte API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v2",
    "paths": {
...
        "/employees": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated search/filter of employees. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Search employees",
                "parameters": [
                    {
                        "description": "Search criteria and pagination",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of employees",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new employee. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Create employee",
                "parameters": [
                    {
                        "description": "Employee data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Employee created",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/employees/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single employee by ID. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Get employee by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Employee not found",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing employee by ID. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Update employee",
                "parameters": [
                    {
                        "description": "Employee data to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee updated",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft-deletes an employee by ID. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Delete employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee deleted",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/employees/{id}/restore": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restores a soft-deleted employee. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Restore employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee restored",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeCreateRequest": {
            "description": "Name, gender, date of birth, photo, license number, years experience, specialty, schedule, optional is_active.",
            "type": "object",
            "required": [
                "date_of_birth",
                "first_name",
                "gender",
                "last_name",
                "license_number",
                "schedule",
                "specialty"
            ],
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "not_specified",
                        "other"
                    ]
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "license_number": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "schedule": {
                    "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeScheduleRequest"
                },
                "specialty": {
                    "type": "string"
                },
                "years_experience": {
                    "type": "integer",
                    "maximum": 60,
                    "minimum": 0
                }
            }
        },
        "clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeResponse": {
            "description": "Employee entity: id, name, gender, date of birth, photo, license, specialty, years experience, is_active, user_id, timestamps.",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "license_number": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "specialty": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "years_experience": {
                    "type": "integer"
                }
            }
        },
        "clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeScheduleRequest": {
            "description": "Day of week, entry/departure time (0-23), optional break window.",
            "type": "object",
            "required": [
                "day",
                "departure_time",
                "entry_time"
            ],
            "properties": {
                "day": {
                    "description": "e.g. \"monday\"",
                    "type": "string"
                },
                "departure_time": {
                    "type": "integer",
                    "maximum": 23,
                    "minimum": 0
                },
                "end_break": {
                    "type": "integer",
                    "maximum": 23,
                    "minimum": 0
                },
                "entry_time": {
                    "description": "0-23",
                    "type": "integer",
                    "maximum": 23,
                    "minimum": 0
                },
                "start_break": {
                    "type": "integer",
                    "maximum": 23,
                    "minimum": 0
                }
            }
        },
        "clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeSearchRequest": {
            "description": "Pagination plus optional specialty and is_active filter.",
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "order_by": {
                    "type": "string",
                    "maxLength": 50
                },
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "page_size": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "sort_direction": {
                    "enum": [
                        "ASC",
                        "DESC"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_page.SortDirection"
                        }
                    ]
                },
                "specialty": {
                    "type": "string"
                }
            }
        },
        "clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeUpdateRequest": {
            "description": "ID (required) plus optional fields and optional schedule.",
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "not_specified",
                        "other"
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "license_number": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "schedule": {
                    "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.EmployeeScheduleRequest"
                },
                "specialty": {
                    "type": "string"
                },
                "years_experience": {
                    "type": "integer",
                    "maximum": 60,
                    "minimum": 0
                }
            }
        },
...
        "clinic-vet-api_internal_shared_http.APIResponse": {
            "description": "Standardized API response structure.",
            "type": "object",
            "properties": {
                "data": {
                    "description": "The data payload of the response for successful requests."
                },
                "error": {
                    "description": "Details of the error if the request was not successful.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.ErrorInfo"
                        }
                    ]
                },
                "message": {
                    "description": "A message providing additional context about the response.",
                    "type": "string"
                },
                "meta": {
                    "description": "Additional metadata for the response, such as pagination info."
                },
                "request_id": {
                    "description": "A unique identifier for the request.",
                    "type": "string"
                },
                "success": {
                    "description": "Indicates whether the request was successful.",
                    "type": "boolean"
                },
                "timestamp": {
                    "description": "The timestamp when the response was generated.",
                    "type": "string"
                }
            }
        },
        "clinic-vet-api_internal_shared_http.ErrorInfo": {
            "description": "Detailed information about an error.",
            "type": "object",
            "properties": {
                "code": {
                    "description": "The error code.",
                    "type": "string"
                },
                "details": {
                    "description": "A map of specific error details, often used for validation errors.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "A descriptive error message.",
                    "type": "string"
                },
                "stack": {
                    "description": "The error stack trace (for development purposes).",
                    "type": "string"
                },
                "type": {
                    "description": "The type of the error.",
                    "type": "string"
                }
            }
        },
        "clinic-vet-api_internal_shared_page.SortDirection": {
            "type": "string",
            "enum": [
                "ASC",
                "DESC"
            ],
            "x-enum-comments": {
                "ASC": "Ascending order",
                "DESC": "Descending order"
            },
            "x-enum-descriptions": [
                "Ascending order",
                "Descending order"
            ],
            "x-enum-varnames": [
                "ASC",
                "DESC"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token. Example: \"Bearer \u0026lt;access_token\u0026gt;\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}